requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
pandas-datareader==0.10.0
yfinance==0.2.31
pydantic==2.5.1
quantconnect-stubs==2.5.13
lean-engine==2.5.13
